fuck

+++ code/datums/preferences.dm
@@ -50,6 +50,7 @@ datum/preferences
 	var/use_wasd = 1
 	var/use_azerty = 0 // do they have an AZERTY keyboard?
 	var/spessman_direction = SOUTH
+	var/PDAcolor = "#6F7961"
 
 	var/job_favorite = null
 	var/list/jobs_med_priority = list()
@@ -97,16 +98,16 @@ datum/preferences
 		//real_name = random_name(src.gender)
 		if (src.gender == MALE)
 			if (first)
-				src.name_first = capitalize(pick(first_names_male))
+				src.name_first = capitalize(pick_string_autokey("names/first_male.txt"))
 			if (middle)
-				src.name_middle = capitalize(pick(first_names_male))
+				src.name_middle = capitalize(pick_string_autokey("names/first_male.txt"))
 		else
 			if (first)
-				src.name_first = capitalize(pick(first_names_female))
+				src.name_first = capitalize(pick_string_autokey("names/first_female.txt"))
 			if (middle)
-				src.name_middle = capitalize(pick(first_names_female))
+				src.name_middle = capitalize(pick_string_autokey("names/first_female.txt"))
 		if (last)
-			src.name_last = capitalize(pick(last_names))
+			src.name_last = capitalize(pick_string_autokey("names/last.txt"))
 		src.real_name = src.name_first + " " + src.name_last
 
 	proc/randomizeLook() // im laze
@@ -215,6 +216,8 @@ datum/preferences
 	var/list/profile_cache
 	var/rebuild_profile
 
+	var/had_cloud = FALSE
+
 	var/list/rebuild_data
 	var/list/data_cache
 
@@ -230,11 +233,11 @@ datum/preferences
 		if (!data_cache)
 			data_cache = list("script" = null,"css" = null,"profile_name" = null,"character_name" = null,"gender" = null,"age_blood" = null,\
 								"bank" = null,"flavortext" = null,"security_note" = null,"medical_note" = null,"occupation" = null,"traits" = null,\
-								"fartsound" = null,"screamsound" = null,"chatsound" = null,"skintone" = null,"eyecolor" = null,"hair_top" = null,"hair_mid" = null,"hair_bottom" = null,\
+								"fartsound" = null,"screamsound" = null,"chatsound" = null,"PDAcolor"=null,"skintone" = null,"eyecolor" = null,"hair_top" = null,"hair_mid" = null,"hair_bottom" = null,\
 								"underwear" = null,"randomize" = null,"font_size" = null,"messages" = null,"hud" = null,"tooltips" = null, "tgui" = null,"popups" = null,"controls" = null,"map"=null)
 			rebuild_data = list("script" = 1,"css" = 1,"profile_name" = 1,"character_name" = 1,"gender" = 1,"age_blood" = 1,\
 								"bank" = 1,"flavortext" = 1,"security_note" = 1,"medical_note" = 1,"occupation" = 1,"traits" = 1,\
-								"fartsound" = 1,"screamsound" = 1,"chatsound" = 1,"skintone" = 1,"eyecolor" = 1,"hair_top" = 1,"hair_mid" = 1,"hair_bottom" = 1,\
+								"fartsound" = 1,"screamsound" = 1,"chatsound" = 1,"PDAcolor" = 1,"skintone" = 1,"eyecolor" = 1,"hair_top" = 1,"hair_mid" = 1,"hair_bottom" = 1,\
 								"underwear" = 1,"randomize" = 1,"font_size" = 1,"messages" = 1,"hud" = 1,"tooltips" = 1, "tgui" = 1, "popups" = 1,"controls" = 1,"map"=1)
 		if (!profile_cache)
 			profile_cache = list()
@@ -251,7 +254,7 @@ datum/preferences
 
 		var/favoriteJob = src.job_favorite ? find_job_in_controller_by_string(src.job_favorite) : ""
 		//mbc is sorry
-		var/chui_toggle_script_jqery_thing = (user && user.client && !user.client.use_chui) ? "<script type='text/javascript' src='[resource("js/jquery.min.js")]'></script>" : ""
+		var/chui_toggle_script_jqery_thing = (user?.client && !user.client.use_chui) ? "<script type='text/javascript' src='[resource("js/jquery.min.js")]'></script>" : ""
 
 		LAGCHECK(LAG_HIGH)
 		//mbc is sorry
@@ -262,12 +265,17 @@ datum/preferences
 		//var/profile_menu[]
 
 		if (user && !IsGuestKey(user.key)) //ZeWaka: Fix for null.key
-			if (rebuild_profile)
+			var/client/client = ismob( user ) ? user.client : user
+
+			if (!client) return // b r u h
+
+			if (rebuild_profile || client.cloud_available() && !had_cloud)
 				rebuild_profile = 0
+				had_cloud = client.cloud_available()
+				rebuild_data["profile_name"] = 1
 				profile_cache.len = 0
 				profile_cache += "<div id='cloudsaves'><strong>Cloud Saves</strong><hr>"
-				var/client/wtf = ismob( user ) ? user.client : user
-				for( var/name in wtf.cloudsaves )
+				for( var/name in client.player.cloudsaves )
 					profile_cache += "<a href='[pref_link]cloudload=[url_encode(name)]'>[html_encode(name)]</a> (<a href='[pref_link]cloudsave=[url_encode(name)]'>Save</a> - <a href='[pref_link]clouddelete=[url_encode(name)]'>Delete</a>)<br>"
 					LAGCHECK(LAG_REALTIME)
 				profile_cache += "<a href='[pref_link]cloudnew=1'>Create new save</a></div>"
@@ -604,6 +612,19 @@ $(function() {
 		</th>
 		<td colspan="2">
 			<a href='[pref_link]voicetype=input'>[AH.voicetype]</a>
+		</td>
+	</tr>"}
+		LAGCHECK(80)
+		if (rebuild_data["PDAcolor"])
+			rebuild_data["PDAcolor"] = 0
+			data_cache["PDAcolor"] = {"
+	<tr>
+		<th>
+			PDA Backlight<span class="info-thing" title="Your character's default PDA background color.">?</span>
+		</th>
+		<td>
+			<a href='[pref_link]PDAcolor=input'>&#9998;</a>
+			<span class='colorbit' style="background-color: [src.PDAcolor];">[src.PDAcolor]</span>
 		</td>
 	</tr>"}
 		LAGCHECK(80)
@@ -1100,7 +1121,7 @@ $(function() {
 			<th>Medium Priority <span class="info-thing" title="Medium Priority Jobs are any jobs you would like to play that aren't your favorite. People with jobs in this category get priority over those who have the same job in their low priority bracket. It's best to put jobs here that you actively enjoy playing and wouldn't mind ending up with if you don't get your favorite.">?</span></th>
 			<th>Low Priority <span class="info-thing" title="Low Priority Jobs are jobs that you don't mind doing. When the game is finding candidates for a job, it will try to fill it with Medium Priority players first, then Low Priority players if there are still free slots.">?</span></th>
 			<th>Unwanted Jobs <span class="info-thing" title="Unwanted Jobs are jobs that you absolutely don't want to have. The game will never give you a job you list here. The 'Staff Assistant' role can't be put here, however, as it's the fallback job if there are no other openings.">?</span></th>
-			<th>Antagonist Roles <span class="info-thing" title="Antagonist roles are randomly chosen when the game starts, before jobs have been allocated. Having an antagonist role disabled means you will never be chosen for it automatically.">?</span></th>
+			<th>Antagonist Roles <span class="info-thing" title="Antagonist roles are randomly chosen when the game starts, before jobs have been allocated. Leaving an antagonist role unchecked means you will never be chosen for it automatically.">?</span></th>
 		</tr>
 		<tr>"}
 
@@ -1303,7 +1324,7 @@ $(function() {
 		return 1
 
 	Topic(href, href_list[])
-		if (usr && usr.client && usr.client.preferences)
+		if (usr?.client?.preferences)
 			if (src == usr.client.preferences)
 				process_link(usr, href_list)
 			else
@@ -1405,9 +1426,9 @@ $(function() {
 
 				if ("random")
 					if (src.gender == MALE)
-						new_name = capitalize(pick(first_names_male) + " " + capitalize(pick(last_names)))
+						new_name = capitalize(pick_string_autokey("names/first_male.txt") + " " + capitalize(pick_string_autokey("names/last.txt")))
 					else
-						new_name = capitalize(pick(first_names_female) + " " + capitalize(pick(last_names)))
+						new_name = capitalize(pick_string_autokey("names/first_female.txt") + " " + capitalize(pick_string_autokey("names/last.txt")))
 					randomizeLook()
 			if (new_name)
 				if (length(new_name) >= 26)
@@ -1442,9 +1463,9 @@ $(function() {
 					new_name = capitalize(new_name)
 				if ("random")
 					if (src.gender == MALE)
-						new_name = capitalize(pick(first_names_male))
+						new_name = capitalize(pick_string_autokey("names/first_male.txt"))
 					else
-						new_name = capitalize(pick(first_names_female))
+						new_name = capitalize(pick_string_autokey("names/first_female.txt"))
 			if (new_name)
 				src.name_first = new_name
 				src.real_name = src.name_first + " " + src.name_last
@@ -1468,9 +1489,9 @@ $(function() {
 					new_name = capitalize(new_name)
 				if ("random")
 					if (src.gender == MALE)
-						new_name = capitalize(pick(first_names_male))
+						new_name = capitalize(pick_string_autokey("names/first_male.txt"))
 					else
-						new_name = capitalize(pick(first_names_female))
+						new_name = capitalize(pick_string_autokey("names/first_female.txt"))
 			src.name_middle = new_name // don't need to check if there is one in case someone wants no middle name I guess
 // -------------------------------------------
 		if (link_tags["last_name"])
@@ -1498,7 +1519,7 @@ $(function() {
 						return
 					new_name = capitalize(new_name)
 				if ("random")
-					new_name = capitalize(pick(last_names))
+					new_name = capitalize(pick_string_autokey("names/last.txt"))
 			if (new_name)
 				src.name_last = new_name
 				src.real_name = src.name_first + " " + src.name_last
@@ -1624,10 +1645,12 @@ $(function() {
 
 				if(new_tone)
 					AH.s_tone = new_tone
+					AH.s_tone_original = new_tone
 			else
 				new_tone = get_standard_skintone(user)
 				if(new_tone)
 					AH.s_tone = new_tone
+					AH.s_tone_original = new_tone
 
 		if (link_tags["underwear_color"])
 			rebuild_data["underwear"] = 1
@@ -1660,7 +1683,7 @@ $(function() {
 
 		if (link_tags["toggle_mentorhelp"])
 			rebuild_data["messages"] = 1
-			if (user && user.client && user.client.is_mentor())
+			if (user?.client?.is_mentor())
 				src.see_mentor_pms = !(src.see_mentor_pms)
 				user.client.set_mentorhelp_visibility(src.see_mentor_pms)
 
@@ -1702,9 +1725,13 @@ $(function() {
 			src.use_azerty = !src.use_azerty
 			src.keybind_prefs_updated(user.client)
 
+		if (link_tags["PDAcolor"])
+			rebuild_data["PDAcolor"] = 1
+			src.PDAcolor = input(usr, "Choose a color", "PDA", src.PDAcolor) as color | null
+
 		if (link_tags["preferred_map"])
 			rebuild_data["map"] = 1
-			src.preferred_map = mapSwitcher.clientSelectMap(usr.client)
+			src.preferred_map = mapSwitcher.clientSelectMap(usr.client,pickable=0)
 
 		if (link_tags["tooltip"])
 			rebuild_data["tooltips"] = 1
@@ -1854,7 +1881,7 @@ $(function() {
 		*/
 
 		if (!isnull(user) && !IsGuestKey(user.key))
-			if (link_tags["cloudsave"] && user.client.cloudsaves[ link_tags["cloudsave"] ])
+			if (link_tags["cloudsave"] && user.client.player.cloudsaves[ link_tags["cloudsave"] ])
 				rebuild_profile = 1
 				var/ret = src.cloudsave_save( user.client, link_tags["cloudsave"] )
 				if( istext( ret ) )
@@ -1863,7 +1890,7 @@ $(function() {
 					boutput( user, "<span class='notice'>Savefile saved!</span>" )
 			else if (link_tags["cloudnew"])
 				rebuild_profile = 1
-				if( user.client.cloudsaves.len >= SAVEFILE_PROFILES_MAX )
+				if( user.client.player.cloudsaves.len >= SAVEFILE_PROFILES_MAX )
 					alert( user, "You have hit your cloud save limit. Please write over an existing save." )
 				else
 					var/newname = input( user, "What would you like to name the save?", "Save Name" ) as text
@@ -1875,14 +1902,14 @@ $(function() {
 							boutput( user, "<span class='alert'>Failed to save savefile: [ret]</span>" )
 						else
 							boutput( user, "<span class='notice'>Savefile saved!</span>" )
-			else if( link_tags["clouddelete"] && user.client.cloudsaves[ link_tags["clouddelete"] ] && alert( user, "Are you sure you want to delete [link_tags["clouddelete"]]?", "Uhm!", "Yes", "No" ) == "Yes" )
+			else if( link_tags["clouddelete"] && user.client.player.cloudsaves[ link_tags["clouddelete"] ] && alert( user, "Are you sure you want to delete [link_tags["clouddelete"]]?", "Uhm!", "Yes", "No" ) == "Yes" )
 				rebuild_profile = 1
 				var/ret = src.cloudsave_delete( user.client, link_tags["clouddelete"] )
 				if( istext( ret ) )
 					boutput( user, "<span class='alert'>Failed to delete savefile: [ret]</span>" )
 				else
 					boutput( user, "<span class='notice'>Savefile deleted!</span>" )
-			else if (link_tags["cloudload"] && user.client.cloudsaves[ link_tags["cloudload"] ])
+			else if (link_tags["cloudload"] && user.client.player.cloudsaves[ link_tags["cloudload"] ])
 				for (var/x in rebuild_data)
 					rebuild_data[x] = 1
 				rebuild_profile = 1
@@ -1920,13 +1947,14 @@ $(function() {
 			AH.customization_third = "None"
 			AH.underwear = "No Underwear"
 
-			AH.customization_first_color = 0
-			AH.customization_second_color = 0
-			AH.customization_third_color = 0
+			AH.customization_first_color = initial(AH.customization_first_color)
+			AH.customization_second_color = initial(AH.customization_second_color)
+			AH.customization_third_color = initial(AH.customization_third_color)
 			AH.e_color = 0
 			AH.u_color = "#FEFEFE"
 
 			AH.s_tone = "#FAD7D0"
+			AH.s_tone_original = "#FAD7D0"
 
 			age = 30
 			pin = null
@@ -1957,6 +1985,7 @@ $(function() {
 			tooltip_option = TOOLTIP_ALWAYS
 			tgui_fancy = TRUE
 			tgui_lock = FALSE
+			PDAcolor = "#6F7961"
 			if (!force_random_names)
 				be_random_name = 0
 			else
@@ -2223,6 +2252,7 @@ var/global/list/facial_hair = list("None" = "none",
 	"Puffy Beard" = "puffbeard",
 	"Long Beard" = "longbeard",
 	"Tramp" = "tramp",
+	"Motley" = "motley",
 	"Eyebrows" = "eyebrows",
 	"Huge Eyebrows" = "thufir")
 
@@ -2285,10 +2315,10 @@ var/global/list/female_screams = list("female", "femalescream1", "femalescream2"
 		H.sound_scream = AH.screamsounds[pick(AH.gender == MALE ? male_screams : female_screams)]
 	if (H && change_name)
 		if (AH.gender == FEMALE)
-			H.real_name = pick(first_names_female)
+			H.real_name = pick_string_autokey("names/first_female.txt")
 		else
-			H.real_name = pick(first_names_male)
-		H.real_name += " [pick(last_names)]"
+			H.real_name = pick_string_autokey("names/first_male.txt")
+		H.real_name += " [pick_string_autokey("names/last.txt")]"
 
 	AH.voicetype = RANDOM_HUMAN_VOICE
 
@@ -2311,6 +2341,7 @@ var/global/list/female_screams = list("female", "femalescream1", "femalescream2"
 		stone = rand(34,-184)
 
 	AH.s_tone = blend_skintone(stone, stone, stone)
+	AH.s_tone_original = AH.s_tone
 
 	if (H)
 		if (H.limbs)
@@ -2388,15 +2419,11 @@ var/global/list/female_screams = list("female", "femalescream1", "femalescream2"
 	if (H && change_age)
 		H.bioHolder.age = rand(20,80)
 
-	if (H && H.organHolder && H.organHolder.head && H.organHolder.head.donor_appearance) // aaaa
+	if (H?.organHolder?.head?.donor_appearance) // aaaa
 		H.organHolder.head.donor_appearance.CopyOther(AH)
 
 	SPAWN_DBG(1 DECI SECOND)
-		AH.UpdateMob()
-		if (H)
-			H.set_face_icon_dirty()
-			H.set_body_icon_dirty()
-
+		H?.update_colorful_parts()
 
 // Generates a real crap checkbox for html toggle links.
 // it sucks but it's a bit more readable i guess.
